refactor(common): hoist shared instant file name bodies
InstantFileNameGeneratorV1 and InstantFileNameGeneratorV2 were
byte-identical except for five members. Move the 33 shared bodies (the
StringUtils.join extension one-liners plus makeDeltaFileName,
getCommitFromCommitFile, makeFileNameAsComplete and
makeFileNameAsInflight) verbatim into a new abstract
BaseInstantFileNameGenerator in org.apache.hudi.common.table.timeline.

V1 and V2 now extend the base and keep only the layout-specific
members: getLayoutVersion, makeRequestedClusteringFileName,
makeInflightClusteringFileName (V1 aliases the replace-commit names,
V2 has dedicated clustering extensions), getFileName(HoodieInstant)
and getFileName(String, HoodieInstant) (V2 emits
requestedTime_completionTime for completed instants). These stay
abstract so a future layout version must state them explicitly.

The InstantFileNameGenerator interface, DefaultInstantFileNameGenerator
and every caller are unchanged. The only text edit inside a retained
body is V1's "1n 0.x" comment typo, now "In 0.x".
